==> HSCA_code/Build_HSCA/HSCA_extended/4_anndata_to_rds.ipynb <==
This notebook saves the relevant information from the **HSCA extended** Anndata file after transfer learning, allowing it to be exported later as an **RDS file** for further analysis of the HSCA extended in **R**.

from scipy import io
import pandas as pd
import os
import anndata
import scanpy as sc
import numpy as np

os.chdir("/data/projects/Bioinformatics/human_skin_cell_atlas/scripts")

**Note:** Cell numbers here differ from the final HSCA extended, as de novo annotation and integration guided by label transfer uncertainties in **R** reveal low-quality cells that were not detectable during standalone preprocessing.

adata = sc.read("./scarches_output/HSCA/SCVI/HSCA_extended/atlas_with_uncerts.h5ad")
adata

os.chdir("/data/projects/Bioinformatics/human single cell skin atlas/scripts/atlas_seurat_output")

!mkdir matrix_files

adata

np.max(adata.X)

np.max(adata.layers["counts"])

with open('matrix_files/barcodes.tsv', 'w') as f:
    for item in adata.obs_names:
        f.write(item + '\n')

with open('matrix_files/features.tsv', 'w') as f:
    for item in ['\t'.join([x,x,'Gene Expression']) for x in adata.var_names]:
        f.write(item + '\n')

io.mmwrite('matrix_files/matrix', adata.layers["counts"].T)

!ls matrix_files/

!gzip matrix_files/*

!ls matrix_files/

adata.obs.to_csv('metadata.csv')

adata.obsm["X_scvi_emb"]

import pandas as pd

# Access to SCVI-Embeddings
scvi_df = pd.DataFrame(
    adata.obsm["X_scvi_emb"],
    index=adata.obs_names
)

scvi_df

scvi_df.to_csv("scvi_embedding.csv")

import pandas as pd

# Extract UMAP from anndata obj
umap_df = pd.DataFrame(
    adata.obsm["X_umap"],
    index=adata.obs_names
)
umap_df.columns = ["UMAP_1", "UMAP_2"]

# Save
umap_df.to_csv("umap_coords.csv")


import pandas as pd

pca_df = pd.DataFrame(
    adata.obsm["X_pca"],
    index=adata.obs_names
)
pca_df.to_csv("pca_embedding.csv")
